+++ FROST-Server.SQLjooq/src/main/java/de/fraunhofer/iosb/ilt/frostserver/persistence/pgjooq/MariaDBExpressionHandler.java
@@ -33,41 +33,39 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- *
  * @author Hylke van der Schaaf
  */
-public class PgExpressionHandler extends ExpressionHandler {
+public class MariaDBExpressionHandler extends ExpressionHandler {
 
-    private static final String ST_GEOM_FROM_EWKT = "ST_GeomFromEWKT(?)";
+    private static final String ST_GeomFromText = "ST_GeomFromText(?)";
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(PgExpressionHandler.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(MariaDBExpressionHandler.class);
 
-    public PgExpressionHandler(CoreSettings settings, QueryBuilder queryBuilder) {
+    public MariaDBExpressionHandler(CoreSettings settings, QueryBuilder queryBuilder) {
         super(settings, queryBuilder);
     }
 
     @Override
     public FieldWrapper visit(LineStringConstant node) {
         Geometry geom = fromGeoJsonConstant(node);
-        return new SimpleFieldWrapper(DSL.field(ST_GEOM_FROM_EWKT, PostGisGeometryBinding.dataType(), geom.asText()));
+        return new SimpleFieldWrapper(DSL.field(ST_GeomFromText, PostGisGeometryBinding.dataType(), geom.asText()));
     }
 
     @Override
     public FieldWrapper visit(PointConstant node) {
         Geometry geom = fromGeoJsonConstant(node);
-        return new SimpleFieldWrapper(DSL.field(ST_GEOM_FROM_EWKT, PostGisGeometryBinding.dataType(), geom.asText()));
+        return new SimpleFieldWrapper(DSL.field(ST_GeomFromText, PostGisGeometryBinding.dataType(), geom.asText()));
     }
 
     @Override
     public FieldWrapper visit(PolygonConstant node) {
         Geometry geom = fromGeoJsonConstant(node);
-        return new SimpleFieldWrapper(DSL.field(ST_GEOM_FROM_EWKT, PostGisGeometryBinding.dataType(), geom.asText()));
+        return new SimpleFieldWrapper(DSL.field(ST_GeomFromText, PostGisGeometryBinding.dataType(), geom.asText()));
     }
 
-    @Override
     public Geometry fromGeoJsonConstant(GeoJsonConstant<? extends GeoJsonObject> node) {
         if (node.getValue().getCrs() == null) {
-            return Wkt.fromWkt("SRID=4326;" + node.getSource());
+            return Wkt.fromWkt(node.getSource());
         }
         return Wkt.fromWkt(node.getSource());
     }