+++ CHANGELOG.md
@@ -20,6 +20,7 @@ FROST-Server version 2.5 and higher requires Java 21. This is because some libra
 * Improved update efficiency by not doing a separate select to get the full updated entity. (georghas)
 * Fixed #2030: NPE when using anonymous access when authentication is enabled on MQTT.
 * Added conformance class to index document for Projects plugin.
+* Apply fine-grained authorisation to MQTT-Create.
 
 
 ## Release version 2.4.0