+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Internal changes & Bugfixes**
 * Added support for spatial queries on MariaDB.
+* Fixed json batch requests without atomicity group all running in the same group.
 
 
 ## Release version 2.4.2