+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
-## Development version 2.4.3
+## Release version 2.4.3
 
 **Internal changes & Bugfixes**
 * Added support for spatial queries on MariaDB.
 * Fixed json batch requests without atomicity group all running in the same group.
 * Avoid use of UPDATE ... RETURNING ... on MariaDB.
+* Updated depencencies.
 
 
 ## Release version 2.4.2